Move token helper logic to redux actions
